+++ services/core-api/app/api/incidents/namespace.py
@@ -6,10 +6,14 @@ from app.api.incidents.resources.mine_incident_followup_types import MineInciden
 from app.api.incidents.resources.mine_incident_determination_types import MineIncidentDeterminationTypeResource
 from app.api.incidents.resources.mine_incident_document_type_codes import MineIncidentDocumentTypeCodeResource
 from app.api.incidents.resources.incidents_resource import IncidentsResource
+from app.api.incidents.resources.mine_incident_notes import MineIncidentNoteResource, MineIncidentNoteListResource
 
 api = Namespace('incidents', description='Incidents actions/options')
 
 api.add_resource(IncidentsResource, '')
+api.add_resource(MineIncidentNoteListResource, '/<string:mine_incident_guid>/notes')
+api.add_resource(MineIncidentNoteResource,
+                 '/<string:mine_incident_guid>/notes/<string:mine_incident_note_guid>')
 api.add_resource(MineIncidentFollowupTypeResource, '/followup-types')
 api.add_resource(MineIncidentDeterminationTypeResource, '/determination-types')
 api.add_resource(MineIncidentStatusCodeResource, '/status-codes')
